Added DELETE process. Bugfixed ssh-ing parts. Created nodes can be cert-logged in with ssh. Some issues with populating hosts and known_hosts and authorized_keys.

+++ spawn.sh
@@ -38,6 +38,10 @@ function helpme() {
     echo "                            Default: 300 seconds.  Required is an integer"
     echo "    -g | --tag            adds a tag to machines. This helps grouping machines in Linode's GUI"
     echo
+    echo "Node removal:"
+    echo "    REQUIRED"
+    echo "    -d | --delete         list of ID nubmers of nodes to be deleted.  Token is also required (see -t )."
+    echo
     echo "Query Linode:"
     echo "    -q | --query-linode   query Linode API for availabe node sizes and datacenter locations"
     echo
@@ -71,7 +75,7 @@ HOSTFILE_TMP=`date +%y%m%d-%H%M%S-etc-hostfile.tmp`
 KNOWNHST_TMP=`date +%y%m%d-%H%M%S-home-ssh-knownhosts.tmp`
 
 ## ARRAYS:
-#  arrays shall not be initiated in bash as they will have a zero-elmement at first index
+#  arrays are commented-out since they shall not be initiated in bash (because following initiation they'd have an undesired zero-elmement)
 #NODES_ARRAY=("")       # host names from input
 #SIZES_ARRAY=("")
 #LINODE_ID_ARRAY=("")
@@ -79,6 +83,7 @@ KNOWNHST_TMP=`date +%y%m%d-%H%M%S-home-ssh-knownhosts.tmp`
 #PRVATE_IP_ARRAY=("")
 #SSH_FINGERPRT_A=("")   # ssh finger print array
 #SSH_PUBLIC_CERT=("")   # node's pub cert
+#IDS_TO_DELETE=("")
 
 ## INTEGERS:
 NODE_LAST_INDEX=-1      # after parsing the input, will contain last node input; for loops. 
@@ -86,11 +91,14 @@ NODE_LAST_INDEX=-1      # after parsing the input, will contain last node input;
 ## BOOLEANS:
 VERBOSE=false
 QUERY=false
+DELETING=false
 
 
 ####################
 # Argument parsing #
 ####################
+
+echo -e "`date +%y%m%d-%H%M`\t${0} $@"  | sed 's/--token\ [0-9a-h]*/--token\ XXXXXXXXXXXXXXXX/1' >> $LOG_FILE
 
 while [[ $# -gt 0 ]]; do
 
@@ -152,6 +160,15 @@ while [[ $# -gt 0 ]]; do
             echo "Verbose."
             shift
             ;;
+        -d | --delete)
+            DELETING=true
+            while [[ ! -z ${2} ]] && [[ ! ${2:0:1} == "-" ]]; do
+                IDS_TO_DELETE+=("${2}")
+                shift
+            done             
+            echo "DELETING..."
+            shift
+            ;;
         -h | --help)
             helpme
             ;;
@@ -163,9 +180,9 @@ done
 
 
 
-######################
-# Intermediate tests #
-######################
+########################
+# Initial common tests #
+########################
 
 # no token? get help
 if [ -z "${TOKEN}" ];  then echo "ERROR: No token provided.  See the help page.";  helpme;  fi
@@ -173,51 +190,77 @@ if $VERBOSE; then echo "<<"$TOKEN">>"; fi
 # TODO: bad token? notify!
 
 
-# ??? one more? : [node_array length>0] and [${#NODES_ARRAY[@]} == ${#SIZES_ARRAY[@]}]  iff its not a query
-# test # for value in "${NODES_ARRAY[@]}"; do echo $value; done
-# test # for value in "${SIZES_ARRAY[@]}"; do echo $value; done
-
-
 NODE_LAST_INDEX=$(( ${#NODES_ARRAY[@]} - 1 ))
-
-
-# in case of incorrect node-size (a.k.a. node-type) array length, populate the array with defaults
-if ! [ ${#NODES_ARRAY[@]} = ${#SIZES_ARRAY[@]} ]; then
-    echo "WARNINIG: Assuming the default nod size/type of: ${TYPE} because \"--sizes\" array is not the same length as \"--nodelist\" array."
-    for ANINDEX in $(seq 0 $NODE_LAST_INDEX); do
-        SIZES_ARRAY+=("${TYPE}")
-    done
-fi
-
-
-echo -e "\n\n" > ${HOSTFILE_TMP}
-if ! [ -f ${HOSTFILE_TMP} ]; then          echo "ERROR: Temporary hostfile ${HOSTFILE_TMP} cannot be written or read.  Fix that, please.";  exit 1;  fi
-echo -e "\n\n" > ${KNOWNHST_TMP}
-if ! [ -f ${KNOWNHST_TMP} ]; then          echo "ERROR: Temporary known_hosts ${KNOWNHST_TMP} cannot be written or read.  Fix that, please.";  exit 1;  fi
-if ! [ -f ${CERT_FILE} ]; then             echo "WARNINIG: Certificate file cannot be read. If you provided no password, you will have troubles logging into your new nodes."; fi
-###if ! command -v sshpass &> /dev/null; then echo "WARNINIG: no sshpass found.  Ssh key is therefore required.  (Or break and install epel-release and sshpass.)";  fi
-
+DELE_LAST_INDEX=$(( ${#IDS_TO_DELETE[@]} - 1 ))
 
 if $VERBOSE; then set -x; fi
 
 
 
-###################
-# Spawn the nodes #
-###################
+################
+# Delete nodes #
+################
+
+if [ ${DELETING} = "true" ]; then
+
+    for ANINDEX in $(seq 0 $DELE_LAST_INDEX); do
+
+        NODE_ID_NUM=${IDS_TO_DELETE[$ANINDEX]}
+
+        # curl generates 1-liner yaml  |  sed splits into lines  |  grep choses label-line only  |  sed then cuts all before label-text and after label-text
+        NODE_NAME=`curl -s -H "Authorization: Bearer $TOKEN" https://api.linode.com/v4/linode/instances/${NODE_ID_NUM} | sed 's/,/\n/g' | grep label | sed 's/^.*:\ "//1' | sed 's/".*//1'`
+
+        echo "-->   == DELETING  id=$NODE_ID_NUM  label=$NODE_NAME... =="
+        read -p "         You sure? (yes/no, default: no) " yn
+
+        case $yn in 
+            yes )
+                OUTPUT=`curl -s -H "Authorization: Bearer $TOKEN" -X DELETE https://api.linode.com/v4/linode/instances/$NODE_ID_NUM`;
+                echo ${OUTPUT};
+                echo -e "`date +%y%m%d-%H%M`\tdeleting ${NODE_ID_NUM}, ${OUTPUT}" >> $LOG_FILE;;
+            * ) echo "Skiped.";
+                echo -e "`date +%y%m%d-%H%M`\tskipping ${NODE_ID_NUM}" >> $LOG_FILE;;
+        esac
+
+    done
+
+    exit 0;
+fi
 
 
-# For npw assume the cert is there.  Requirement easily met.
-#    if ! [ -f ${CERT_FILE} ]; then
-#        # no cert in payload
-#    else
-#        # payload includes cert   
-#    fi
+#######################
+# More specific tests #
+#######################
 
-# assuming the CERT_FILE is a file holding a valid public part of a local certificate that will later be used for ssh authentication
+if [ ${DELETING} = "false" ]; then
+
+    # in case of incorrect node-size (a.k.a. node-type) array length, populate the array with defaults
+    if ! [ ${#NODES_ARRAY[@]} = ${#SIZES_ARRAY[@]} ]; then
+        echo "WARNINIG: Assuming the default nod size/type of: ${TYPE} because \"--sizes\" array is not the same length as \"--nodelist\" array."
+        for ANINDEX in $(seq 0 $NODE_LAST_INDEX); do
+            SIZES_ARRAY+=("${TYPE}")
+        done
+    fi
+
+    echo -e "\n\n" > ${HOSTFILE_TMP}
+    if ! [ -f ${HOSTFILE_TMP} ]; then          echo "ERROR: Temporary hostfile ${HOSTFILE_TMP} cannot be written or read.  Fix that, please.";  exit 1;  fi
+    echo -e "\n\n" > ${KNOWNHST_TMP}
+    if ! [ -f ${KNOWNHST_TMP} ]; then          echo "ERROR: Temporary known_hosts ${KNOWNHST_TMP} cannot be written or read.  Fix that, please.";  exit 1;  fi
+    if ! [ -f ${CERT_FILE} ]; then             echo "WARNINIG: Certificate file cannot be read. If you provided no password, you will have troubles logging into your new nodes."; fi
+
+fi
+
+
+
+
+
+###############
+# Spawn nodes #
+###############
 
 CERT_ESCAPED=`cat $CERT_FILE | sed 's:\/:\\\/:g'`
-TOKEN_STR=`cat ~/token2311master`
+
+echo -e "LINODE_ID\tNODENAME\tLI_PUB_IP\tLI_PRV_IP\tNODESIZE\tIMAGE\tREGIO" >> $LOG_FILE
 
 for ANINDEX in $(seq 0 $NODE_LAST_INDEX); do
 
@@ -228,7 +271,7 @@ for ANINDEX in $(seq 0 $NODE_LAST_INDEX); do
 
     YAML_PAYLOAD="{\"type\": \""$NODESIZE"\", \"region\": \""$REGIO"\", \"image\": \""$IMAGE"\", \"root_pass\": \""$PASSW"\", \"label\": \""$NODENAME"\", \"authorized_keys\": [\""$CERT_ESCAPED"\"] }"
 
-    OUTPUT=`curl --progress-bar -X POST https://api.linode.com/v4/linode/instances -H "Authorization: Bearer $TOKEN_STR" -H "Content-type: application/json" -d "$YAML_PAYLOAD"`
+    OUTPUT=`curl --progress-bar -X POST https://api.linode.com/v4/linode/instances -H "Authorization: Bearer $TOKEN" -H "Content-type: application/json" -d "$YAML_PAYLOAD"`
 
     LINODE_ID=`echo $OUTPUT | grep -E -o "id\":\ [0-9]+" | sed 's/id\":\ //g'`
     LINODE_ID_ARRAY+=($LINODE_ID)
@@ -251,7 +294,7 @@ for ANINDEX in $(seq 0 $NODE_LAST_INDEX); do
 
     # add a hostsfile entry for that machine
     echo -e "${LI_PRV_IP}\t${NODENAME}" >> ${HOSTFILE_TMP}
-
+    echo "created " $LINODE_ID $NODENAME $$LI_PUB_IP $LI_PRV_IP >> ${LOG_FILE}
     echo "Server pas: $PASSW"
 
 done
@@ -294,30 +337,27 @@ for ANINDEX in $(seq 0 $NODE_LAST_INDEX); do
 
     OUTPUT=`ssh-keyscan -t ecdsa ${IP_PUBL} 2>&1 | grep ecdsa`      # not the best practice to write functions (despite inline!) TWICE (1)
     while [ -z "$OUTPUT" ]; do
-        echo "wait some more for ecdsa signature of $IP_PUBL ($NODNAME)..."
+        echo " \- wait some more for ecdsa signature of $IP_PUBL ($NODNAME)..."
         sleep 5
-        # ssh-keyscan -t ecdsa ${IP_PUBL}
         OUTPUT=`ssh-keyscan -t ecdsa ${IP_PUBL} 2>&1 | grep ecdsa`  # not the best practice to write functions (despite inline!) TWICE (2)
     done
     SSH_FP_COMPLETE="${NODNAME},${IP_PRIV},${OUTPUT}"           # IP_PUB is included anyway in the ssh-keyscan output, at the beginning
     SSH_FINGERPRT_A+=(${SSH_FP_COMPLETE})
 
-    echo ${SSH_FP_COMPLETE}
+    echo ${SSH_FP_COMPLETE}                                     #
     echo ${SSH_FP_COMPLETE} >> ${KNOWNHST_TMP}                  # ecdsa fingerprint: to OPERATOR CONSOLE, and to TEMP_FILE
-    echo ${NODNAME},${OUTPUT} >> ~/.ssh/known_hosts             #                    and to local known_hosts file
+    echo ${NODNAME},${OUTPUT} >> ~/.ssh/known_hosts             #                    and to LOCAL known_hosts file
     
-    # ssh-rsa public key for authorized_keys
+    # generate ssh key-pair on nodes.  BEWARE: the entrophy on all newly created nodes may be "the same", with little difference,
+    # based on the image used for the node, which brings SECURITY QUESTIONS about the quality of such ssh keys
+############################# PROBLEM HERE
+############################# PROBLEM HERE
+############################# PROBLEM HERE
+    ssh root@${IP_PUBL} "ssh-keygen -f ~/.ssh/id_rsa -q -P ''"
+    # and record public keys
+    SERVER_PUB_CERT=`ssh root@${IP_PUBL} "cat ~/.ssh/id_rsa.pub"`
+    SSH_PUBLIC_CERT+=${SERVER_PUB_CERT}                         # pub keys for saving later in use in ~/.ssh/authorized_keys of all nodes, so all-to-all have the connectivity
     
-    # 1 - operator's console ssh key for key-authenticatoin
-    ###echo "Copying ssh id to authorized_keys of $NODENAME"
-    #sshpass -p ${PASSW} ssh-copy-id -i ${CERT_FILE} root@${LI_PUB_IP}    
-    ###sshpass -p ${PASSW} ssh-copy-id root@${LI_PUB_IP}
-    
-    # 2 - generate and obtain server's public key
-    ssh root@${LI_PUB_IP} "ssh-keygen -f ~/.ssh/id_rsa -q -P ''"
-    # ssh root@${LI_PUB_IP} "/root/cat .ssh/id_rsa.pub"
-    SERVER_PUB_CERT=`ssh root@${LI_PUB_IP} "cat ~/.ssh/id_rsa.pub"`
-    SSH_PUBLIC_CERT+=${SERVER_PUB_CERT}                         # for saving later in all nodes, so all-to-all have the connectivity
     echo "Sever $NODENAME .ssh/rsa_id.pub is: $SERVER_PUB_CERT"
     
 done
@@ -330,17 +370,11 @@ done
 # create hostsfile entries
 # create known_hosts entries
 #
-# add all to known_hosts
-# copy-ssh-key
+# copy-ssh-key # i.e. add all to known_hosts
 # 
-echo;echo;echo;echo;
-echo "Let's review first!"
-echo;echo;echo
 
 cat ${HOSTFILE_TMP}
 cat ${KNOWNHST_TMP}
-
-exit
 
 for ANINDEX in $(seq 0 $NODE_LAST_INDEX); do
 
@@ -348,16 +382,24 @@ for ANINDEX in $(seq 0 $NODE_LAST_INDEX); do
     NODNAME=${NODES_ARRAY[$ANINDEX]}
 
     # propagate hostsfile
-    cat ${HOSTFILE_TMP} | ssh root@${LI_PUB_IP} "dd >> /etc/hosts"
+    cat ${HOSTFILE_TMP} | ssh root@${IP_PUBL} "dd >> /etc/hosts"
 
     # propagate .ssh/known_hosts
-    cat ${KNOWNHST_TMP} | ssh root@${LI_PUB_IP} "dd >> /root/.ssh/known_hosts"
+    cat ${KNOWNHST_TMP} | ssh root@${IP_PUBL} "dd >> /root/.ssh/known_hosts"
     
-    # propagate autorized_keys
-    ssh root@${LI_PUB_IP} "mkdir -p ~/.ssh/; cat >> ~/.ssh/authorized_keys; chmod 700 ~/.ssh/; chmod 600 ~/.ssh/authorized_keys"
-    
+    # create autorized_keys
+    # cat >> auth_k # is in order to add a possibly missing tailing eol.
+    ssh root@${IP_PUBL} "mkdir -p ~/.ssh/; cat >> ~/.ssh/authorized_keys; chmod 700 ~/.ssh/; chmod 600 ~/.ssh/authorized_keys"
     # populate .ssh/authorized_keys between servers in the environment
-    
+
+    for SUBINDEX in $(seq 0 $NODE_LAST_INDEX); do
+        if [ "$SUBINDEX" != "$ANINDEX" ]; then                  # don't add your own cert to your own auth_k - why create mess, just don't.
+        
+            CERT_TO_ADD=${SERVER_PUB_CERT[$SUBINDEX]}
+            ssh root@${IP_PUBL} "echo ${CERT_TO_ADD} >> /root/.ssh/authorized_keys"
+        
+        fi
+    done
     
     # yum install object storage tools
     # LOL, maybe: ssh help or cowasy greeting on remote node installed
